add [palette] section to skin INI files, for reference in syntax files
instead of choosing specific colors in a skin, and finding that some
of the particular colors in a syntax file don't display legibly; we
specify color palette to be used in the skin file, and also a new
syntax section that assigns colors (either terminal colors, or named
colors from the palette section) to syntax classes.  then we change
the syntax files to use syntax classes instead of bare colors, and
now syntax highlighting color schemes match the skin palette.

i've switched over a hand full of the syntaxes i use to the new
named classes scheme, and added default syntax colors to all the
shipped skins.  I haven't checked that files are legible with each
combination of syntax file, skin file, and terminal color palette.
but there were plenty of illegible combinations before, so this is
no worse - better for me in fact, because now I get the same colors
for comments/strings/constants in all of the languages I use, plus
I can change them all to match my customized skin in one place
rather than keeping copies of each file.